# lib/api-zod/src/generated/types/createWorkoutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWorkoutBodyWeightUnit } from "./createWorkoutBodyWeightUnit";

export interface CreateWorkoutBody {
  exerciseId: string;
  exerciseName: string;
  detectedWeightKg?: number | null;
  weightUnit: CreateWorkoutBodyWeightUnit;
  notes?: string | null;
}