# lib/api-zod/src/generated/types/detectWeightBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DetectWeightBodyUnit } from "./detectWeightBodyUnit";

export interface DetectWeightBody {
  /** Base64-encoded image data */
  imageBase64: string;
  unit?: DetectWeightBodyUnit;
}