# lib/api-zod/src/generated/types/detectWeightResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DetectWeightResultConfidence } from "./detectWeightResultConfidence";

export interface DetectWeightResult {
  detected: boolean;
  weightKg?: number | null;
  weightLbs?: number | null;
  confidence?: DetectWeightResultConfidence;
  description?: string | null;
}