# lib/api-zod/src/generated/types/workoutSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutSessionWeightUnit } from "./workoutSessionWeightUnit";
import type { WorkoutSet } from "./workoutSet";

export interface WorkoutSession {
  id: string;
  exerciseId: string;
  exerciseName: string;
  detectedWeightKg?: number | null;
  weightUnit: WorkoutSessionWeightUnit;
  startedAt: Date;
  completedAt?: Date | null;
  sets: WorkoutSet[];
  totalVolume?: number | null;
  notes?: string | null;
}